Move/rename Attribute::Content
- It’s now an inner class of Core::ItemFactory
