+++ lib/munge.rb
@@ -8,7 +8,6 @@ require "rack"
 require "tilt"
 
 require "munge/version"
-require "munge/attribute/content"
 require "munge/item"
 require "munge/helper"
 require "munge/helper/find"
@@ -20,6 +19,7 @@ require "munge/core/config"
 require "munge/core/renderer"
 require "munge/core/router"
 require "munge/core/item_factory"
+require "munge/core/item_factory/content_parser"
 require "munge/core/source"
 require "munge/core/write"
 require "munge/core/alterant"
